addObservation 1st step
