+++ src/app/health_records/about_patient/[patient_id]/components/clinicVisitsDashboard.jsx
@@ -22,7 +22,8 @@ import {
   DropdownMenuCheckboxItem,
 } from "@/components/ui/dropdown-menu";
 import ViewClinicVisit from "./sub_components/viewClinicVisit";
-import AddObservation from "./sub_components/sub_sub_components/addObservation";
+// import AddObservation from "./sub_components/sub_sub_components/addObservation";
+import AddClinicVisit from "./sub_components/addClinicVisit";
 import * as React from "react";
 import BackButton from "./sub_components/BackButton";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
@@ -317,7 +318,7 @@ export default function ClinicVisits({ patientId }) {
         </>
       ) : currentPage === 10 ? (
         <>
-          <AddObservation
+          <AddClinicVisit
             currentPage={currentPage}
             setCurrentPage={setCurrentPage}
             patientId={patientId}
